Refactor into two services
